requirements.txt
streamlit==1.44.1
pytubefix==8.12.3
pydub==0.25.1
elevenlabs==1.56.1
python-dotenv==1.1.0
openai==1.75.0